Refactoring: utils_load is available
